CHROMIUM: thunderbolt: Configure link after lane bonding is enabled
During testing it was noticed that the link is not properly restored
after the domain exits sleep if the link configured bits are set before
lane bonding is enabled. The USB4 spec does not say in which order these
need to be set but setting link configured afterwards makes the link
restoration work so we do so instead.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=b:160191015
TEST=Observed no regressions with S0iX without an attached device.

TOREVERT=Once the upstream version of this patch lands in maintainer
tree, this patch should be reverted.

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: Ie8febe8a9dc39af799c34a5e161fe68b23e6b3e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2275612
Tested-by: Dana Alkattan <dana.alkattan@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>