FROMGIT: mfd: Add support for the MediaTek MT6358 PMIC
This adds support for the MediaTek MT6358 PMIC. This is a
multifunction device with the following sub modules:

- Regulator
- RTC
- Codec
- Interrupt

It is interfaced to the host controller using SPI interface
by a proprietary hardware called PMIC wrapper or pwrap.
MT6358 MFD is a child device of the pwrap.

Signed-off-by: Hsin-Hsiung Wang <hsin-hsiung.wang@mediatek.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 2b91c28f2abd9471aac4aa6cd8de7896ef469153
 git://git.kernel.org/pub/scm/linux/kernel/git/lee/mfd.git ib-mfd-power-rtc-5.8)

BUG=b:126006378
TEST=Build kukui kernel

Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Change-Id: I82525ede0374d73cf08d35ff98e97eba9d6673a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2214649
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Claire Chang <tientzu@chromium.org>