BACKPORT: FROMLIST: drm/amd/display: Use cursor locking to prevent flip delays
[Why]
Current locking scheme for cursor can result in a flip missing
its vsync, deferring it for one or more vsyncs.  Result is a
potential for stuttering when cursor is moved.

[How]
Use cursor update lock so that flips are not blocked while cursor
is being programmed.

Signed-off-by: Aric Cyr <aric.cyr@amd.com>
Reviewed-by: Nicholas Kazlauskas <Nicholas.Kazlauskas@amd.com>
Acked-by: Aurabindo Pillai <aurabindo.pillai@amd.com>
(am from https://patchwork.freedesktop.org/patch/362717/)

BUG=b:166558695
BUG=b:166228317
TEST=build & boot

Change-Id: Ie7ce27cf91e5dec99aa746b3e3d986b0b1c4a68d
Signed-off-by: Ikshwaku Chauhan <ikshwaku.chauhan@amd.corp-partner.google.com>
Signed-off-by: Deepak Sharma <deepak.sharma@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2381329
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Tested-by: Shirish S <shirish.s@amd.com>