UPSTREAM: drm/i915/gt: Keep a no-frills swappable copy of the default context state
We need to keep the default context state around to instantiate new
contexts (aka golden rendercontext), and we also keep it pinned while
the engine is active so that we can quickly reset a hanging context.
However, the default contexts are large enough to merit keeping in
swappable memory as opposed to kernel memory, so we store them inside
shmemfs. Currently, we use the normal GEM objects to create the default
context image, but we can throw away all but the shmemfs file.

This greatly simplifies the tricky power management code which wants to
run underneath the normal GT locking, and we definitely do not want to
use any high level objects that may appear to recurse back into the GT.
Though perhaps the primary advantage of the complex GEM object is that
we aggressively cache the mapping, but here we are recreating the
vm_area everytime time we unpark. At the worst, we add a lightweight
cache, but first find a microbenchmark that is impacted.

Having started to create some utility functions to make working with
shmemfs objects easier, we can start putting them to wider use, where
GEM objects are overkill, such as storing persistent error state.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matthew Auld <matthew.auld@intel.com>
Cc: Ramalingam C <ramalingam.c@intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200429172429.6054-1-chris@chris-wilson.co.uk
(cherry picked from commit be1cb55a07bfc528d826962f1e421a44ed5f8311)

BUG=b:168669388
TEST=Check for GPU frequency samples above 1100MHz.
     - Login as guest and launch WebGL Aquarium with 30,000 fish.
     - Monitor frequency statistics.
       `socwatch -f cpu -f gfx -t 120`
       alternatively
       `while true; do cat /sys/class/drm/card0/gt_act_freq_mhz; sleep 0.1; done`

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I3acd715b049b1409045a792f10643704c2163623
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2414870
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>