UPSTREAM: drm/i915/gt: Always enable busy-stats for execlists
In the near future, we will utilize the busy-stats on each engine to
approximate the C0 cycles of each, and use that as an input to a manual
RPS mechanism. That entails having busy-stats always enabled and so we
can remove the enable/disable routines and simplify the pmu setup. As a
consequence of always having the stats enabled, we can also show the
current active time via sysfs/engine/xcs/active_time_ns.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200429205446.3259-1-chris@chris-wilson.co.uk
(cherry picked from commit 426d0073fb6d1a9513978cea4c9e8396f3721fba)

BUG=b:168669388
TEST=Check for GPU frequency samples above 1100MHz.
     - Login as guest and launch WebGL Aquarium with 30,000 fish.
     - Monitor frequency statistics.
       `socwatch -f cpu -f gfx -t 120`
       alternatively
       `while true; do cat /sys/class/drm/card0/gt_act_freq_mhz; sleep 0.1; done`

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: Ibf945f36ab917f87901da1671da62e1f675cd8ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2414871
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>