UPSTREAM: drm/i915/gt: Switch to manual evaluation of RPS
As with the realisation for soft-rc6, we respond to idling the engines
within microseconds, far faster than the response times for HW RC6 and
RPS. Furthermore, our fast parking upon idle, prevents HW RPS from
running for many desktop workloads, as the RPS evaluation intervals are
on the order of tens of milliseconds, but the typical workload is just a
couple of milliseconds, but yet we still need to determine the best
frequency for user latency versus power.

Recognising that the HW evaluation intervals are a poor fit, and that
they were deprecated [in bspec at least] from gen10, start to wean
ourselves off them and replace the EI with a timer and our accurate
busy-stats. The principle benefit of manually evaluating RPS intervals
is that we can be more responsive for better performance and powersaving
for both spiky workloads and steady-state.

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/1698
Fixes: 98479ada421a ("drm/i915/gt: Treat idling as a RPS downclock event")
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Cc: Andi Shyti <andi.shyti@intel.com>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200429205446.3259-4-chris@chris-wilson.co.uk
(cherry picked from commit 36d516be867ce3657401ab26b05a39a20f56fa13)

BUG=b:168669388
TEST=Check for GPU frequency samples above 1100MHz.
     - Login as guest and launch WebGL Aquarium with 30,000 fish.
     - Monitor frequency statistics.
       `socwatch -f cpu -f gfx -t 120`
       alternatively
       `while true; do cat /sys/class/drm/card0/gt_act_freq_mhz; sleep 0.1; done`

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I1d5257ce98d454d0745abd20c058be1992172ad7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2414874
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>