FROMGIT: net: ipa: do not cache channel state
It is possible for a GSI channel's state to be changed as a result
of an action by a different execution environment.  Specifically,
the modem is able to issue a GSI generic command that causes a state
change on a GSI channel associated with the AP.

A channel's state only needs to be known when a channel is allocated
or deallocaed, started or stopped, or reset.  So there is little
value in caching the state anyway.

Stop recording a copy of the channel's last known state, and instead
fetch the true state from hardware whenever it's needed.  In such
cases, *do* record the state in a local variable, in case an error
message reports it (so the value reported is the value seen).

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit a2003b303875b41542bad1c2e81800fdd4c27c29
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git/ master)

BUG=b:74020136
TEST=Boot modem on trogdor

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I58c0d902817ce3f8c07e77969b1360b65c736e61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2194544
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>