CHROMIUM: arm64: dts: qcom: Invert polarity of eDP lane 0 on sc7180-lazor
It appears swapped on the cable.  We can use the "lane-polarities"
property in the bridge chip to account for this.

NOTE: at the moment the driver doesn't support this property, but it
will soon.  We can still land the dts bits ahead of time since they're
already specified in the bindings.

BUG=b:155686212
TEST=Lazor noise gone

Change-Id: I157a56b84cb34f4f5a855c1b734cb348ff07bd65
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2180901
Reviewed-by: Stephen Boyd <swboyd@chromium.org>