BACKPORT: UPSTREAM: Bluetooth: Replace BT_DBG with bt_dev_dbg for management support
The majority of management interaction are based on a controller index
and have a hci_dev associated with it. So use bt_dev_dbg to have a clean
way of indentifying the controller the debug message belongs to.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
(cherry picked from commit 181d695352305cc52a49c151a1c3370376e54887)

Conflicts:
	net/bluetooth/mgmt.c

BUG=none
TEST=build

Signed-off-by: Zhengping Jiang <jiangzp@google.com>