FROMGIT: net: ipa: use tag process on modem crash
One part of recovering from a modem crash is performing a "tag
sequence" of several IPA immediate commands, to clear the hardware
pipeline.  The sequence ends with a data transfer request on the
command endpoint (which is not otherwise done).  Unfortunately,
attempting to do the data transfer led to a hang, so that request
plus two other commands were commented out.

The previous commit fixes the bug that was causing that hang.  And
with that bug fixed we can properly issue the tag sequence when the
modem crashes, to return the hardware to a known state.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 2c4bb8093c3bfebc8113c683a5de83c64d43514c
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net.git master)

BUG=b:74020136
TEST=Boot modem on trogdor

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I398a014ba9098628d796280f036e2dcc88a5c038
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2204458
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>